Crate vkxml

Structs

Bitmask
Definition that specifies a bitmask, using a particular existing type. Contents are the same as typedef. The possible bits can be defined by a specific, referenced enumeration.

Command
Defines a single Vulkan entrypoint.

Commands
Defines a list of Vulkan entrypoints.

Constant
C constant

Constants

Define
C #define statement.

Definitions
List of all definitions used by Vulkan

Enumeration

EnumerationDeclaration
Declares the name of an enumeration. The members and base type of the enumeration are defined elsewhere.

Enums

Extension
Specifies the types exposed/forbidden by an extension

ExtensionConstant
Defines a constant exported by an extension. These are not enumerators.

ExtensionEnum
Declares a new enumerator that is added to an existing enumerator.

ExtensionSpecification

Extensions

ExternalSync
Describes instances of implicit external synchronizations. These cannot be captured by the parameters themselves.

Feature
Defines the base API for a Vulkan specification. Represents a specific major/minor version of Vulkan.

FeatureSpecification

Features

Field
A typed field used as a member in a struct/union, or as a function parameter.

FunctionPointer
Defines a function pointer.

Handle
Handle created using VK_DEFINE_HANDLE or VK_DEFINE_NON_DISPATCHABLE_HANDLE.

Include
Definition that represents a file being included, which is named. The name attribute does double-duty, in that it also includes text.

NamedIdentifier

Range
Represents a range of enumerators.

Reference
Definition that references a type that potentially comes from an include file.

Registry
The root of vk_new.xml

ReturnType

Struct
Definition that represents a data structure. Contains a list of members and optional validation information.

Tag
A single author for Vulkan

Tags
List of authors

Typedef
Definition that defines a new name for an existing type.

Union
Defines a union, where the object's value is only one of its members.

VendorId

VendorIds
List of identifiers for vendors.

Enums

ArrayType
C array type

DefinitionsElement

EnumerationElement

EnumerationPurpose

EnumsElement

ExtensionElement

ExtensionSpecificationElement

ExtensionType
Specifies the kind of extension

FeatureElement

FeatureReference

HandleType
Whether the handle is dispatchable or not.

IncludeStyle
Specifies whether to use "" or <> around the include name.

Pipeline

ReferenceType

RegistryElement

Renderpass

StructElement

Type Definitions

ArraySize
If the array is "static", ArraySize specifies an integer literal.

CExpression
Text that is an expression in C.

CommaSeparatedBooleans
Comma-separated boolean values.

CommaSeparatedIdentifiers
Comma-separated identifiers values. Can be empty.

HexadecimalNumber
Recognize a hexadecimal number. Not prefixed with "0x"

Identifier
Should be a valid C/C++ identifier.

Notation
C comment